requirements.txt
Django==4.0
pylint==2.6.0
pylint-django==2.3.0
autopep8==1.5.4
nose==1.3.7
django-nose==1.4.7
Unidecode==1.3.4
django-crispy-forms==1.14.0
django-allauth==0.51.0

